refactor: players 일급컬렉션 생성
