+++ src/main/java/blackjack/domain/entry/Players.java
@@ -1,40 +1,21 @@
 package blackjack.domain.entry;
 
 import blackjack.domain.PlayerOutcome;
-import blackjack.domain.card.Deck;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 public class Players {
-    private final Dealer dealer;
     private final List<Player> players;
 
-    public Players(Dealer dealer, List<Player> players) {
-        this.dealer = dealer;
+    public Players(List<Player> players) {
         this.players = players;
     }
 
-    public boolean isDealerHit(Deck deck) {
-        if (dealer.shouldHaveMoreCard()) {
-            dealer.putCard(deck.draw());
-            return true;
-        }
-        return false;
-    }
-
-    public Map<PlayerOutcome, List<Player>> getGameResult() {
+    public Map<PlayerOutcome, List<Player>> match(Dealer dealer) {
         return players.stream()
                 .collect(Collectors.groupingBy(player -> player.match(dealer)));
-    }
-
-    public List<Participant> getParticipant() {
-        List<Participant> participants = new ArrayList<>();
-        participants.add(dealer);
-        participants.addAll(players);
-        return participants;
     }
 
     public List<Player> getPlayers() {
